split compiler-plugins tests
